requirements.txt
python-tk==3.8.4
mysql-connector-python==8.0.20
 Pillow==7.2.0 
